Concluida a base do CRUD do aplicativo

+++ app/schema.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS "produtos" (
     "codigo" INTEGER PRIMARY KEY,
     "nome" TEXT NOT NULL,
     "descricao" TEXT NOT NULL,
-    "preco" NUMERIC NOT NULL,
+    "preco" REAL NOT NULL,
     "imagem" TEXT NOT NULL
 );
